move examples + examples generation methods
